requirements.txt

requests
pandas
numpy
ta
python-dotenv
PyYAML